add animatecounter.js to handle the counter increasing effect
