--- frontend/js/components/animateNumber.js
@@ -1,1 +1,0 @@
-export default function () {}
