+++ static/css/style.css
@@ -2133,6 +2133,10 @@ body:not(.js-home-follower) .section-projects.dark, .js-home-follower .section-p
 
 @media only screen and (min-width: 1024px) { .section-mixed-content-row h2 { margin-bottom: 20px; } }
 
+.section-mixed-content-row .mobile-background { z-index: 1; visibility: hidden; position: absolute; top: 0; left: 0; right: 0; bottom: 0; background-size: cover; background-position: center; -webkit-transition: -webkit-transform 100ms ease-out; transition: -webkit-transform 100ms ease-out; transition: transform 100ms ease-out; transition: transform 100ms ease-out, -webkit-transform 100ms ease-out; }
+
+@media only screen and (max-width: 1023px) { .section-mixed-content-row .mobile-background { visibility: visible; } }
+
 .section-mixed-content-row .background { position: absolute; top: 0; left: 0; right: 0; bottom: 0; background-size: cover; background-position: center; -webkit-transition: -webkit-transform 100ms ease-out; transition: -webkit-transform 100ms ease-out; transition: transform 100ms ease-out; transition: transform 100ms ease-out, -webkit-transform 100ms ease-out; }
 
 .section-mixed-content-row.has-bg-image { padding: 60px 0 20px !important; }
@@ -2143,7 +2147,7 @@ body:not(.js-home-follower) .section-projects.dark, .js-home-follower .section-p
 
 @media only screen and (min-width: 1024px) { .section-mixed-content-row.has-bg-image .content { margin-top: -80px; margin-bottom: 340px; } }
 
-.section-mixed-content-row.has-bg-image::before { display: block; position: absolute; top: 0; left: 0; right: 0; height: 130px; background: -webkit-gradient(linear, left top, left bottom, from(#ffffff), to(rgba(255, 255, 255, 0))); background: linear-gradient(to bottom, #ffffff, rgba(255, 255, 255, 0)); content: ''; z-index: 1; }
+.section-mixed-content-row.has-bg-image::before { display: block; position: absolute; top: 0; left: 0; right: 0; height: 130px; background: -webkit-gradient(linear, left top, left bottom, from(#ffffff), to(rgba(255, 255, 255, 0))); background: linear-gradient(to bottom, #ffffff, rgba(255, 255, 255, 0)); content: ""; z-index: 1; }
 
 @media only screen and (min-width: 1024px) { .section-mixed-content-row.has-bg-image::before { height: 170px; } }
