+++ static/js/script.js
@@ -98,6 +98,18 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var core
 
 /***/ }),
 
+/***/ "./frontend/js/components/animatecounters.js":
+/*!***************************************************!*\
+  !*** ./frontend/js/components/animatecounters.js ***!
+  \***************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! core-js/modules/es.object.to-string */ \"./node_modules/core-js/modules/es.object.to-string.js\");\n/* harmony import */ var core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var core_js_modules_es_regexp_exec__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! core-js/modules/es.regexp.exec */ \"./node_modules/core-js/modules/es.regexp.exec.js\");\n/* harmony import */ var core_js_modules_es_regexp_exec__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_regexp_exec__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! core-js/modules/es.regexp.to-string */ \"./node_modules/core-js/modules/es.regexp.to-string.js\");\n/* harmony import */ var core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var core_js_modules_es_string_replace__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! core-js/modules/es.string.replace */ \"./node_modules/core-js/modules/es.string.replace.js\");\n/* harmony import */ var core_js_modules_es_string_replace__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_string_replace__WEBPACK_IMPORTED_MODULE_3__);\n\n\n\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (function () {\n  // Get Countries List from <script> data block.\n  var $counter = $(\".counter\");\n  if (!$counter.length) return; //\n\n  function commaSeparateNumber(val) {\n    while (/(\\d+)(\\d{3})/.test(val.toString())) {\n      val = val.toString().replace(/(\\d+)(\\d{3})/, \"$1\" + \",\" + \"$2\");\n    }\n\n    return val;\n  }\n\n  $(\".counter\").each(function () {\n    var $this = $(this),\n        countTo = $this.attr(\"data-count\");\n    $({\n      countNum: $this.text()\n    }).animate({\n      countNum: countTo\n    }, {\n      duration: 2000,\n      easing: \"swing\",\n      step: function step() {\n        $this.text(commaSeparateNumber(Math.floor(this.countNum)));\n      },\n      complete: function complete() {\n        $this.text(commaSeparateNumber(this.countNum));\n      }\n    });\n  });\n});\n\n//# sourceURL=webpack:///./frontend/js/components/animatecounters.js?");
+
+/***/ }),
+
 /***/ "./frontend/js/components/article-list.js":
 /*!************************************************!*\
   !*** ./frontend/js/components/article-list.js ***!
@@ -394,7 +406,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var core
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! core-js/modules/es.object.to-string */ \"./node_modules/core-js/modules/es.object.to-string.js\");\n/* harmony import */ var core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var core_js_modules_es_regexp_exec__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! core-js/modules/es.regexp.exec */ \"./node_modules/core-js/modules/es.regexp.exec.js\");\n/* harmony import */ var core_js_modules_es_regexp_exec__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_regexp_exec__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! core-js/modules/es.regexp.to-string */ \"./node_modules/core-js/modules/es.regexp.to-string.js\");\n/* harmony import */ var core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var core_js_modules_es_string_replace__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! core-js/modules/es.string.replace */ \"./node_modules/core-js/modules/es.string.replace.js\");\n/* harmony import */ var core_js_modules_es_string_replace__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_string_replace__WEBPACK_IMPORTED_MODULE_3__);\n/* harmony import */ var remove_focus_outline__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! remove-focus-outline */ \"./node_modules/remove-focus-outline/outline.js\");\n/* harmony import */ var remove_focus_outline__WEBPACK_IMPORTED_MODULE_4___default = /*#__PURE__*/__webpack_require__.n(remove_focus_outline__WEBPACK_IMPORTED_MODULE_4__);\n/* harmony import */ var swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! swiper/dist/js/swiper.esm.js */ \"./node_modules/swiper/dist/js/swiper.esm.js\");\n/* harmony import */ var vanilla_lazyload__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! vanilla-lazyload */ \"./node_modules/vanilla-lazyload/dist/lazyload.min.js\");\n/* harmony import */ var vanilla_lazyload__WEBPACK_IMPORTED_MODULE_6___default = /*#__PURE__*/__webpack_require__.n(vanilla_lazyload__WEBPACK_IMPORTED_MODULE_6__);\n/* harmony import */ var smooth_scroll__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! smooth-scroll */ \"./node_modules/smooth-scroll/dist/smooth-scroll.polyfills.min.js\");\n/* harmony import */ var smooth_scroll__WEBPACK_IMPORTED_MODULE_7___default = /*#__PURE__*/__webpack_require__.n(smooth_scroll__WEBPACK_IMPORTED_MODULE_7__);\n/* harmony import */ var aos__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! aos */ \"./node_modules/aos/dist/aos.js\");\n/* harmony import */ var aos__WEBPACK_IMPORTED_MODULE_8___default = /*#__PURE__*/__webpack_require__.n(aos__WEBPACK_IMPORTED_MODULE_8__);\n/* harmony import */ var js_cookie__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! js-cookie */ \"./node_modules/js-cookie/src/js.cookie.js\");\n/* harmony import */ var js_cookie__WEBPACK_IMPORTED_MODULE_9___default = /*#__PURE__*/__webpack_require__.n(js_cookie__WEBPACK_IMPORTED_MODULE_9__);\n/* harmony import */ var _components_polyfills__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! ./components/polyfills */ \"./frontend/js/components/polyfills.js\");\n/* harmony import */ var _components_donation_tabs__WEBPACK_IMPORTED_MODULE_11__ = __webpack_require__(/*! ./components/donation-tabs */ \"./frontend/js/components/donation-tabs.js\");\n/* harmony import */ var _components_dollar_handles__WEBPACK_IMPORTED_MODULE_12__ = __webpack_require__(/*! ./components/dollar-handles */ \"./frontend/js/components/dollar-handles.js\");\n/* harmony import */ var _components_toggles__WEBPACK_IMPORTED_MODULE_13__ = __webpack_require__(/*! ./components/toggles */ \"./frontend/js/components/toggles.js\");\n/* harmony import */ var _components_en_petition_form__WEBPACK_IMPORTED_MODULE_14__ = __webpack_require__(/*! ./components/en-petition-form */ \"./frontend/js/components/en-petition-form.js\");\n/* harmony import */ var _components_anchor_menu__WEBPACK_IMPORTED_MODULE_15__ = __webpack_require__(/*! ./components/anchor-menu */ \"./frontend/js/components/anchor-menu.js\");\n/* harmony import */ var _components_countries_menu__WEBPACK_IMPORTED_MODULE_16__ = __webpack_require__(/*! ./components/countries-menu */ \"./frontend/js/components/countries-menu.js\");\n/* harmony import */ var _components_swipers__WEBPACK_IMPORTED_MODULE_17__ = __webpack_require__(/*! ./components/swipers */ \"./frontend/js/components/swipers.js\");\n/* harmony import */ var _components_follow_unfollow__WEBPACK_IMPORTED_MODULE_18__ = __webpack_require__(/*! ./components/follow-unfollow */ \"./frontend/js/components/follow-unfollow.js\");\n/* harmony import */ var _components_share_to_fixed__WEBPACK_IMPORTED_MODULE_19__ = __webpack_require__(/*! ./components/share-to-fixed */ \"./frontend/js/components/share-to-fixed.js\");\n/* harmony import */ var _components_tips__WEBPACK_IMPORTED_MODULE_20__ = __webpack_require__(/*! ./components/tips */ \"./frontend/js/components/tips.js\");\n/* harmony import */ var _components_back_links__WEBPACK_IMPORTED_MODULE_21__ = __webpack_require__(/*! ./components/back-links */ \"./frontend/js/components/back-links.js\");\n/* harmony import */ var _components_parallax__WEBPACK_IMPORTED_MODULE_22__ = __webpack_require__(/*! ./components/parallax */ \"./frontend/js/components/parallax.js\");\n/* harmony import */ var _components_scroll_way__WEBPACK_IMPORTED_MODULE_23__ = __webpack_require__(/*! ./components/scroll-way */ \"./frontend/js/components/scroll-way.js\");\n/* harmony import */ var _components_article_list__WEBPACK_IMPORTED_MODULE_24__ = __webpack_require__(/*! ./components/article-list */ \"./frontend/js/components/article-list.js\");\n/* harmony import */ var _components_facebook_video__WEBPACK_IMPORTED_MODULE_25__ = __webpack_require__(/*! ./components/facebook-video */ \"./frontend/js/components/facebook-video.js\");\n/* harmony import */ var _components_dynamic_search__WEBPACK_IMPORTED_MODULE_26__ = __webpack_require__(/*! ./components/dynamic-search */ \"./frontend/js/components/dynamic-search.js\");\n/* harmony import */ var _components_ugc__WEBPACK_IMPORTED_MODULE_27__ = __webpack_require__(/*! ./components/ugc */ \"./frontend/js/components/ugc.js\");\n/* harmony import */ var _petition_thankyou__WEBPACK_IMPORTED_MODULE_28__ = __webpack_require__(/*! ./petition-thankyou */ \"./frontend/js/petition-thankyou.js\");\n/* harmony import */ var jquery_appear_original__WEBPACK_IMPORTED_MODULE_29__ = __webpack_require__(/*! jquery-appear-original */ \"./node_modules/jquery-appear-original/index.js\");\n/* harmony import */ var jquery_appear_original__WEBPACK_IMPORTED_MODULE_29___default = /*#__PURE__*/__webpack_require__.n(jquery_appear_original__WEBPACK_IMPORTED_MODULE_29__);\n\n\n\n\n\n\n\n\n\n\n\n\n // import followingTags from './components/following-tags';\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n // import donationPage from './donation-page';\n\nswiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_5__[\"Swiper\"].use([swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_5__[\"Navigation\"], swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_5__[\"Pagination\"], swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_5__[\"Scrollbar\"], swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_5__[\"Controller\"], swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_5__[\"Autoplay\"]]);\nObject(_components_polyfills__WEBPACK_IMPORTED_MODULE_10__[\"default\"])();\ndocument.addEventListener(\"DOMContentLoaded\", function (event) {\n  new vanilla_lazyload__WEBPACK_IMPORTED_MODULE_6___default.a({\n    elements_selector: \".lazy\",\n    threshold: 400\n  });\n  var Scroll = new smooth_scroll__WEBPACK_IMPORTED_MODULE_7___default.a('a[href*=\"#\"]', {\n    header: \"body > header\",\n    offset: 40,\n    updateURL: false\n  }); // check cookies consent\n\n  var cookieConsentGp = js_cookie__WEBPACK_IMPORTED_MODULE_9___default.a.get(\"greenpeace\");\n\n  if (typeof cookieConsentGp !== \"undefined\" && cookieConsentGp !== 2 || !cookieConsentGp) {\n    var cookieMessage = document.querySelector(\"#set-cookie\");\n    if (cookieMessage) cookieMessage.display = \"block\";\n  }\n  /* Components */\n\n\n  Object(_components_donation_tabs__WEBPACK_IMPORTED_MODULE_11__[\"default\"])();\n  Object(_components_dollar_handles__WEBPACK_IMPORTED_MODULE_12__[\"default\"])(); //followingTags();\n\n  Object(_components_toggles__WEBPACK_IMPORTED_MODULE_13__[\"default\"])();\n  Object(_components_en_petition_form__WEBPACK_IMPORTED_MODULE_14__[\"default\"])();\n  Object(_components_anchor_menu__WEBPACK_IMPORTED_MODULE_15__[\"default\"])();\n  Object(_components_countries_menu__WEBPACK_IMPORTED_MODULE_16__[\"default\"])();\n  Object(_components_swipers__WEBPACK_IMPORTED_MODULE_17__[\"default\"])(swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_5__[\"Swiper\"]);\n  setTimeout(_components_follow_unfollow__WEBPACK_IMPORTED_MODULE_18__[\"default\"], 0);\n  Object(_components_share_to_fixed__WEBPACK_IMPORTED_MODULE_19__[\"default\"])();\n  Object(_components_tips__WEBPACK_IMPORTED_MODULE_20__[\"default\"])();\n  Object(_components_back_links__WEBPACK_IMPORTED_MODULE_21__[\"default\"])();\n  Object(_components_parallax__WEBPACK_IMPORTED_MODULE_22__[\"default\"])();\n  Object(_components_scroll_way__WEBPACK_IMPORTED_MODULE_23__[\"default\"])();\n  Object(_components_article_list__WEBPACK_IMPORTED_MODULE_24__[\"default\"])();\n  Object(_components_facebook_video__WEBPACK_IMPORTED_MODULE_25__[\"default\"])();\n  Object(_components_dynamic_search__WEBPACK_IMPORTED_MODULE_26__[\"default\"])();\n  Object(_components_ugc__WEBPACK_IMPORTED_MODULE_27__[\"default\"])();\n  aos__WEBPACK_IMPORTED_MODULE_8___default.a.init({\n    easing: \"ease-in-out-cubic\",\n    once: true,\n    disable: \"phone\"\n  });\n  /**\n   * Open/close country offices megamenu (not even putting it in a separate module)\n   */\n\n  var $countryOfficesDropdown = $(\"header .dropdown\");\n  $countryOfficesDropdown.on(\"click\", function (e) {\n    e.stopPropagation();\n  });\n  $countryOfficesDropdown.children(\"button\").on(\"click\", function () {\n    $countryOfficesDropdown.toggleClass(\"is-open\");\n  });\n  $(window).click(function () {\n    $countryOfficesDropdown.removeClass(\"is-open\");\n  });\n  /* Page specific functionality */\n\n  Object(_petition_thankyou__WEBPACK_IMPORTED_MODULE_28__[\"default\"])(Scroll);\n});\n/* GPEA extended functions */\n\n\ndocument.addEventListener(\"DOMContentLoaded\", function (event) {\n  // animate at the first time\n  animatecounters(); //\n\n  $(\".counter\").appear();\n  $(document.body).on(\"appear\", \".counter\", function (e) {\n    if (!$(this).hasClass(\"appear\")) {\n      animatecounters();\n      $(this).addClass(\"appear\");\n      console.log(\"appear\");\n    }\n  });\n});\n\nfunction animatecounters() {\n  // Get Countries List from <script> data block.\n  var $counter = $(\".counter\");\n  if (!$counter.length) return; //\n\n  function commaSeparateNumber(val) {\n    while (/(\\d+)(\\d{3})/.test(val.toString())) {\n      val = val.toString().replace(/(\\d+)(\\d{3})/, \"$1\" + \",\" + \"$2\");\n    }\n\n    return val;\n  }\n\n  $(\".counter\").each(function () {\n    var $this = $(this),\n        countTo = $this.attr(\"data-count\");\n    $({\n      countNum: $this.text()\n    }).animate({\n      countNum: countTo\n    }, {\n      duration: 1000,\n      easing: \"swing\",\n      step: function step() {\n        $this.text(commaSeparateNumber(Math.floor(this.countNum)));\n      },\n      complete: function complete() {\n        $this.text(commaSeparateNumber(this.countNum));\n      }\n    });\n  });\n}\n\n//# sourceURL=webpack:///./frontend/js/index.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var remove_focus_outline__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! remove-focus-outline */ \"./node_modules/remove-focus-outline/outline.js\");\n/* harmony import */ var remove_focus_outline__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(remove_focus_outline__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! swiper/dist/js/swiper.esm.js */ \"./node_modules/swiper/dist/js/swiper.esm.js\");\n/* harmony import */ var vanilla_lazyload__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! vanilla-lazyload */ \"./node_modules/vanilla-lazyload/dist/lazyload.min.js\");\n/* harmony import */ var vanilla_lazyload__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(vanilla_lazyload__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var smooth_scroll__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! smooth-scroll */ \"./node_modules/smooth-scroll/dist/smooth-scroll.polyfills.min.js\");\n/* harmony import */ var smooth_scroll__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(smooth_scroll__WEBPACK_IMPORTED_MODULE_3__);\n/* harmony import */ var aos__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! aos */ \"./node_modules/aos/dist/aos.js\");\n/* harmony import */ var aos__WEBPACK_IMPORTED_MODULE_4___default = /*#__PURE__*/__webpack_require__.n(aos__WEBPACK_IMPORTED_MODULE_4__);\n/* harmony import */ var js_cookie__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! js-cookie */ \"./node_modules/js-cookie/src/js.cookie.js\");\n/* harmony import */ var js_cookie__WEBPACK_IMPORTED_MODULE_5___default = /*#__PURE__*/__webpack_require__.n(js_cookie__WEBPACK_IMPORTED_MODULE_5__);\n/* harmony import */ var _components_polyfills__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./components/polyfills */ \"./frontend/js/components/polyfills.js\");\n/* harmony import */ var _components_donation_tabs__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./components/donation-tabs */ \"./frontend/js/components/donation-tabs.js\");\n/* harmony import */ var _components_dollar_handles__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! ./components/dollar-handles */ \"./frontend/js/components/dollar-handles.js\");\n/* harmony import */ var _components_toggles__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! ./components/toggles */ \"./frontend/js/components/toggles.js\");\n/* harmony import */ var _components_en_petition_form__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! ./components/en-petition-form */ \"./frontend/js/components/en-petition-form.js\");\n/* harmony import */ var _components_anchor_menu__WEBPACK_IMPORTED_MODULE_11__ = __webpack_require__(/*! ./components/anchor-menu */ \"./frontend/js/components/anchor-menu.js\");\n/* harmony import */ var _components_countries_menu__WEBPACK_IMPORTED_MODULE_12__ = __webpack_require__(/*! ./components/countries-menu */ \"./frontend/js/components/countries-menu.js\");\n/* harmony import */ var _components_swipers__WEBPACK_IMPORTED_MODULE_13__ = __webpack_require__(/*! ./components/swipers */ \"./frontend/js/components/swipers.js\");\n/* harmony import */ var _components_follow_unfollow__WEBPACK_IMPORTED_MODULE_14__ = __webpack_require__(/*! ./components/follow-unfollow */ \"./frontend/js/components/follow-unfollow.js\");\n/* harmony import */ var _components_share_to_fixed__WEBPACK_IMPORTED_MODULE_15__ = __webpack_require__(/*! ./components/share-to-fixed */ \"./frontend/js/components/share-to-fixed.js\");\n/* harmony import */ var _components_tips__WEBPACK_IMPORTED_MODULE_16__ = __webpack_require__(/*! ./components/tips */ \"./frontend/js/components/tips.js\");\n/* harmony import */ var _components_back_links__WEBPACK_IMPORTED_MODULE_17__ = __webpack_require__(/*! ./components/back-links */ \"./frontend/js/components/back-links.js\");\n/* harmony import */ var _components_parallax__WEBPACK_IMPORTED_MODULE_18__ = __webpack_require__(/*! ./components/parallax */ \"./frontend/js/components/parallax.js\");\n/* harmony import */ var _components_scroll_way__WEBPACK_IMPORTED_MODULE_19__ = __webpack_require__(/*! ./components/scroll-way */ \"./frontend/js/components/scroll-way.js\");\n/* harmony import */ var _components_article_list__WEBPACK_IMPORTED_MODULE_20__ = __webpack_require__(/*! ./components/article-list */ \"./frontend/js/components/article-list.js\");\n/* harmony import */ var _components_facebook_video__WEBPACK_IMPORTED_MODULE_21__ = __webpack_require__(/*! ./components/facebook-video */ \"./frontend/js/components/facebook-video.js\");\n/* harmony import */ var _components_dynamic_search__WEBPACK_IMPORTED_MODULE_22__ = __webpack_require__(/*! ./components/dynamic-search */ \"./frontend/js/components/dynamic-search.js\");\n/* harmony import */ var _components_ugc__WEBPACK_IMPORTED_MODULE_23__ = __webpack_require__(/*! ./components/ugc */ \"./frontend/js/components/ugc.js\");\n/* harmony import */ var _petition_thankyou__WEBPACK_IMPORTED_MODULE_24__ = __webpack_require__(/*! ./petition-thankyou */ \"./frontend/js/petition-thankyou.js\");\n/* harmony import */ var jquery__WEBPACK_IMPORTED_MODULE_25__ = __webpack_require__(/*! jquery */ \"./node_modules/jquery/dist/jquery.js\");\n/* harmony import */ var jquery__WEBPACK_IMPORTED_MODULE_25___default = /*#__PURE__*/__webpack_require__.n(jquery__WEBPACK_IMPORTED_MODULE_25__);\n/* harmony import */ var jquery_appear_original__WEBPACK_IMPORTED_MODULE_26__ = __webpack_require__(/*! jquery-appear-original */ \"./node_modules/jquery-appear-original/index.js\");\n/* harmony import */ var jquery_appear_original__WEBPACK_IMPORTED_MODULE_26___default = /*#__PURE__*/__webpack_require__.n(jquery_appear_original__WEBPACK_IMPORTED_MODULE_26__);\n/* harmony import */ var _components_animatecounters__WEBPACK_IMPORTED_MODULE_27__ = __webpack_require__(/*! ./components/animatecounters */ \"./frontend/js/components/animatecounters.js\");\n\n\n\n\n\n\n\n\n // import followingTags from './components/following-tags';\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n // import donationPage from './donation-page';\n\nswiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_1__[\"Swiper\"].use([swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_1__[\"Navigation\"], swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_1__[\"Pagination\"], swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_1__[\"Scrollbar\"], swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_1__[\"Controller\"], swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_1__[\"Autoplay\"]]);\nObject(_components_polyfills__WEBPACK_IMPORTED_MODULE_6__[\"default\"])();\ndocument.addEventListener(\"DOMContentLoaded\", function (event) {\n  new vanilla_lazyload__WEBPACK_IMPORTED_MODULE_2___default.a({\n    elements_selector: \".lazy\",\n    threshold: 400\n  });\n  var Scroll = new smooth_scroll__WEBPACK_IMPORTED_MODULE_3___default.a('a[href*=\"#\"]', {\n    header: \"body > header\",\n    offset: 40,\n    updateURL: false\n  }); // check cookies consent\n\n  var cookieConsentGp = js_cookie__WEBPACK_IMPORTED_MODULE_5___default.a.get(\"greenpeace\");\n\n  if (typeof cookieConsentGp !== \"undefined\" && cookieConsentGp !== 2 || !cookieConsentGp) {\n    var cookieMessage = document.querySelector(\"#set-cookie\");\n    if (cookieMessage) cookieMessage.display = \"block\";\n  }\n  /* Components */\n\n\n  Object(_components_donation_tabs__WEBPACK_IMPORTED_MODULE_7__[\"default\"])();\n  Object(_components_dollar_handles__WEBPACK_IMPORTED_MODULE_8__[\"default\"])(); //followingTags();\n\n  Object(_components_toggles__WEBPACK_IMPORTED_MODULE_9__[\"default\"])();\n  Object(_components_en_petition_form__WEBPACK_IMPORTED_MODULE_10__[\"default\"])();\n  Object(_components_anchor_menu__WEBPACK_IMPORTED_MODULE_11__[\"default\"])();\n  Object(_components_countries_menu__WEBPACK_IMPORTED_MODULE_12__[\"default\"])();\n  Object(_components_swipers__WEBPACK_IMPORTED_MODULE_13__[\"default\"])(swiper_dist_js_swiper_esm_js__WEBPACK_IMPORTED_MODULE_1__[\"Swiper\"]);\n  setTimeout(_components_follow_unfollow__WEBPACK_IMPORTED_MODULE_14__[\"default\"], 0);\n  Object(_components_share_to_fixed__WEBPACK_IMPORTED_MODULE_15__[\"default\"])();\n  Object(_components_tips__WEBPACK_IMPORTED_MODULE_16__[\"default\"])();\n  Object(_components_back_links__WEBPACK_IMPORTED_MODULE_17__[\"default\"])();\n  Object(_components_parallax__WEBPACK_IMPORTED_MODULE_18__[\"default\"])();\n  Object(_components_scroll_way__WEBPACK_IMPORTED_MODULE_19__[\"default\"])();\n  Object(_components_article_list__WEBPACK_IMPORTED_MODULE_20__[\"default\"])();\n  Object(_components_facebook_video__WEBPACK_IMPORTED_MODULE_21__[\"default\"])();\n  Object(_components_dynamic_search__WEBPACK_IMPORTED_MODULE_22__[\"default\"])();\n  Object(_components_ugc__WEBPACK_IMPORTED_MODULE_23__[\"default\"])();\n  aos__WEBPACK_IMPORTED_MODULE_4___default.a.init({\n    easing: \"ease-in-out-cubic\",\n    once: true,\n    disable: \"phone\"\n  });\n  /**\n   * Open/close country offices megamenu (not even putting it in a separate module)\n   */\n\n  var $countryOfficesDropdown = jquery__WEBPACK_IMPORTED_MODULE_25___default()(\"header .dropdown\");\n  $countryOfficesDropdown.on(\"click\", function (e) {\n    e.stopPropagation();\n  });\n  $countryOfficesDropdown.children(\"button\").on(\"click\", function () {\n    $countryOfficesDropdown.toggleClass(\"is-open\");\n  });\n  jquery__WEBPACK_IMPORTED_MODULE_25___default()(window).click(function () {\n    $countryOfficesDropdown.removeClass(\"is-open\");\n  });\n  /* Page specific functionality */\n\n  Object(_petition_thankyou__WEBPACK_IMPORTED_MODULE_24__[\"default\"])(Scroll);\n});\n/* GPEA extended functions */\n\n\n\n\ndocument.addEventListener(\"DOMContentLoaded\", function (event) {\n  var $counter = jquery__WEBPACK_IMPORTED_MODULE_25___default()(\".counter\");\n  if (!$counter.length) return; // animate at the first time\n\n  Object(_components_animatecounters__WEBPACK_IMPORTED_MODULE_27__[\"default\"])();\n  $counter.appear();\n  jquery__WEBPACK_IMPORTED_MODULE_25___default()(document.body).on(\"appear\", \".counter\", function (e) {\n    if (!jquery__WEBPACK_IMPORTED_MODULE_25___default()(this).hasClass(\"appear\")) {\n      Object(_components_animatecounters__WEBPACK_IMPORTED_MODULE_27__[\"default\"])();\n      jquery__WEBPACK_IMPORTED_MODULE_25___default()(this).addClass(\"appear\");\n    }\n  });\n});\n\n//# sourceURL=webpack:///./frontend/js/index.js?");
 
 /***/ }),
